requirements.txt
fastapi==0.110.0
uvicorn==0.24.0.post1
aiohttp==3.10.11
python-binance==1.0.29
pandas==2.2.2
yfinance==0.2.37
